Consult app env for data and metadata column types
This is ugly, but it gets the job done. We create two module
attributes that get the `:data_column_data_type` and
`:meta_data_column_data_type` from the application environment
configuration. If one of the keys is missing, the corresponding data
type will default to "bytea".

At this time, there is no way for serializer modules using the
serializer behaviour to discern between data and metadata. For this
reason, the data column and metadata column must be the same type;
otherwise, the serializer may fail.