Add a JSONB serializer for tests
The JsonbSerializer implements the EventStore.Serializer
behaviour. Its `serialize/1` function will convert structs to maps
with string keys, which seems to be required for the tests to pass. It
might be a good idea to investigate changing the tests to support this
case; however, the consistency of the tests is more important.

The `deserialize/2` function reverses a map with string keys into the
the struct given in its configuration.