Use PostgreSQL's NOTIFY/LISTEN for event pub/sub
Using PostgreSQL for pub/sub removes need for EventStore to form a cluster of nodes. It can be used on a single node, or multiple nodes, but they don't need to be directly connected. Instead the database will be used to distribute events. This simplifies running EventStore on a cluster.