Close database connection when subscription fails to acquire exclusive lock
Connection will be opened for each subscribe request. This reduces the number of permanently open connections to the database.

To subscribe to a stream a database connection is opened and an attempt to acquire exclusive access to the subscription is made. On success, the connection remains open for ack's, but on failure it is immediately closed.