Use a shared Postgres connection for all subscriptions
A single connection, with no pooling, is used by subscriptions to allow session based Postgres advisory locks. A subscription attempts to acquire an advisory lock before subscribing to events.

This ensures only a single subscription runs, regardless of whether the EventStore is run on multiple nodes, even when not connected together to form a cluster.