Use two 32-bit key values for advisory locks
The namespace key is derived from the unique `oid` value for the `subscriptions` table. The `oid` is unique within a database and differs for identically named tables defined in different schemas and on repeat table definitions.

This change aims to prevent lock collision with application level advisory lock usage and other libraries using Postgres advisory locks. Now there is a 1 in 2,147,483,647 chance of colliding with other locks.